Adding config telemetry for feature flags.
