+++ config/feature_flags.go
@@ -100,29 +100,6 @@ func featureFlagsFromMap(featuresMap map[string]string, baseFeatureFlags model.F
 	return baseFeatureFlags
 }
 
-// featureFlagsToMap returns the feature flags as a map[string]string
-// Supports boolean and string feature flags.
-func featureFlagsToMap(featureFlags *model.FeatureFlags) map[string]string {
-	refStructVal := reflect.ValueOf(*featureFlags)
-	refStructType := reflect.TypeOf(*featureFlags)
-	ret := make(map[string]string)
-	for i := 0; i < refStructVal.NumField(); i++ {
-		refFieldVal := refStructVal.Field(i)
-		refFieldType := refStructType.Field(i)
-		if !refFieldVal.IsValid() {
-			continue
-		}
-		switch refFieldType.Type.Kind() {
-		case reflect.Bool:
-			ret[refFieldType.Name] = strconv.FormatBool(refFieldVal.Bool())
-		default:
-			ret[refFieldType.Name] = refFieldVal.String()
-		}
-	}
-
-	return ret
-}
-
 func getStructFields(s interface{}) []string {
 	structType := reflect.TypeOf(s)
 	fieldNames := make([]string, 0, structType.NumField())
