+++ services/telemetry/telemetry_test.go
@@ -368,6 +368,7 @@ func TestRudderTelemetry(t *testing.T) {
 			TrackConfigExperimental,
 			TrackConfigAnalytics,
 			TrackConfigPlugin,
+			TrackFeatureFlags,
 			TrackActivity,
 			TrackServer,
 			TrackConfigMessageExport,
@@ -410,6 +411,7 @@ func TestRudderTelemetry(t *testing.T) {
 			TrackConfigExperimental,
 			TrackConfigAnalytics,
 			TrackConfigPlugin,
+			TrackFeatureFlags,
 			TrackActivity,
 			TrackServer,
 			TrackConfigMessageExport,
